优化记忆策略

+++ src/main/java/com/agent/hopaw/task/LongTermMemoryTaskHandler.java
@@ -2,10 +2,12 @@ package com.agent.hopaw.task;
 
 import com.agent.hopaw.constant.AiModelCallSourceEnum;
 import com.agent.hopaw.constant.DefaultUser;
+import com.agent.hopaw.constant.LongTermMemoryTypeEnum;
 import com.agent.hopaw.mapper.AgentMapper;
 import com.agent.hopaw.mapper.ChatMemoryMapper;
 import com.agent.hopaw.model.Agent;
 import com.agent.hopaw.model.ChatMemory;
+import com.agent.hopaw.model.LongTermMemory;
 import com.agent.hopaw.model.ScheduledTask;
 import com.agent.hopaw.service.*;
 import com.agent.hopaw.util.InvocationParametersWrapper;
@@ -13,6 +15,7 @@ import com.alibaba.fastjson2.JSON;
 import dev.langchain4j.data.message.*;
 import dev.langchain4j.invocation.InvocationParameters;
 import dev.langchain4j.model.chat.ChatModel;
+import dev.langchain4j.model.chat.request.ChatRequestParameters;
 import dev.langchain4j.service.AiServices;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
@@ -21,6 +24,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -65,48 +69,21 @@ public class LongTermMemoryTaskHandler implements TaskHandler {
     }
 
     private String getConfig(String key, String defaultValue) {
-        var config = sysConfigService.getByKey(key);
-        return config != null ? config.getConfigValue() : defaultValue;
+        return sysConfigService.getValueByKey(key, defaultValue);
     }
 
-    private void processMemoryForIdentity(Agent agent,String userId) {
-
-        //已标记清理的消息
-        List<ChatMemory> cleanedMessages = chatMemoryMapper.findByAgentIdAndUserIdAndStatus(agent.getId(), userId,1);
-        List<ChatMemory> cleanedMessages1 = chatMemoryMapper.findByAgentIdAndUserIdAndStatus(agent.getId(), userId,2);
-        if(!cleanedMessages1.isEmpty()){
-            cleanedMessages.addAll(cleanedMessages1);
-            cleanedMessages=cleanedMessages.stream().sorted(Comparator.comparing(ChatMemory::getCreateTime)).collect(Collectors.toList());
-        }
-        if (cleanedMessages.isEmpty()){
-            return;
-        }
-        int batchSize = agent.getMaxMemoryRecords()/2;
-
-        if(batchSize<=0){
-            batchSize=5;
-        }
-        LocalDateTime latestTime = cleanedMessages.stream()
-                .map(ChatMemory::getCreateTime)
-                .filter(Objects::nonNull)
-                .max(LocalDateTime::compareTo)
-                .orElse(LocalDateTime.now());
-        LocalDateTime fiveMinutesAgo = LocalDateTime.now().minusMinutes(5);
-        if (cleanedMessages.size() < batchSize && latestTime.isAfter(fiveMinutesAgo)) {
-            return;
-        }
-
+    private String buildMessageSummary(List<ChatMemory> cleanedMessages ) {
         StringBuilder conversationBuilder = new StringBuilder();
         for (ChatMemory chat : cleanedMessages) {
             if (chat == null || chat.getMessageJson() == null) {
                 continue;
             }
-            
+
             ChatMessage message = ChatMessageDeserializer.messageFromJson(chat.getMessageJson());
             if (message == null) {
                 continue;
             }
-            
+
             if (message instanceof dev.langchain4j.data.message.UserMessage) {
                 dev.langchain4j.data.message.UserMessage userMessage = ((dev.langchain4j.data.message.UserMessage) message);
 
@@ -180,13 +157,37 @@ public class LongTermMemoryTaskHandler implements TaskHandler {
             }
             conversationBuilder.append("\n");
         }
+        return conversationBuilder.toString();
+    }
 
+    private void processMemoryForIdentity(Agent agent,String userId) {
+        //已标记清理的消息
+        List<ChatMemory> cleanedMessages = chatMemoryMapper.findByAgentIdAndUserIdInStatus(agent.getId(), userId,Arrays.asList(1,2));
+        if (cleanedMessages.isEmpty()){
+            return;
+        }
+        int batchSize = agent.getMaxMemoryRecords()/2;
+        if(batchSize<=0){
+            batchSize=5;
+        }
+        LocalDateTime latestTime = cleanedMessages.stream()
+                .map(ChatMemory::getCreateTime)
+                .filter(Objects::nonNull)
+                .max(LocalDateTime::compareTo)
+                .orElse(LocalDateTime.now());
+        LocalDateTime fiveMinutesAgo = LocalDateTime.now().minusMinutes(5);
+        if (cleanedMessages.size() < batchSize && latestTime.isAfter(fiveMinutesAgo)) {
+            return;
+        }
+        //这是新消息
+        String newConversation = buildMessageSummary(cleanedMessages);
         String agentIdStr = String.valueOf(agent.getId());
-        String newConversation = conversationBuilder.toString();
-        String existingMemory = longTermMemoryService.getMemoryTree(agentIdStr,userId);
 
-        String memory = buildMemorySummary(existingMemory, newConversation);
-        boolean handle = handle(agentIdStr,userId, memory);
+        //现有记忆
+        List<LongTermMemory> longTermMemories = longTermMemoryService.getRecentMemoriesByAgentIdAndUserId(agentIdStr, userId);
+        String content = buildContent(longTermMemories, newConversation);
+
+        boolean handle = handle(agentIdStr,userId, content);
         if (handle) {
             chatMemoryMapper.deleteByIds(cleanedMessages.stream().map(ChatMemory::getId).toList());
         }
@@ -217,7 +218,10 @@ public class LongTermMemoryTaskHandler implements TaskHandler {
                     .tools(Arrays.asList(longTermMemoryService))
                     .build();
             logger.info("开始汇总记忆 \n {}", content);
-            String result = assistant.chat(content,invocationParametersWrapper.getParameters());
+            ChatRequestParameters chatRequestParameters=ChatRequestParameters.builder()
+                    .temperature(0.1)
+                    .build();
+            String result = assistant.chat(content,chatRequestParameters,invocationParametersWrapper.getParameters());
             logger.info("记忆汇总完毕：{}", result);
             return true;
         }catch (Exception ex){
@@ -245,43 +249,42 @@ public class LongTermMemoryTaskHandler implements TaskHandler {
         String customPrompt = getConfig("memory_prompt", "");
         if (customPrompt.isBlank()) {
             customPrompt="你是一个记忆整理助手。善于根据聊天记录提取关键的用户记忆信息。" +
-                    "请根据内容总结出用户的关键记忆信息，并按以下格式进行分类，分类不够可以自己添加，但是分类要精简：\n" +
+                    "请根据现有记忆和新会话总结出用户的关键记忆信息，要严格按以下要求进行分类整理：\n" +
                     "========\n" +
-                    "1,基础档案\n" +
-                    "个人特质、地域作息、性格、身份角色、核心标签、敏感雷区等\n" +
-                    "2,工作职场\n" +
-                    "岗位业务、负责项目、技术 / 专业栈、协作习惯、工作痛点、目标规划、常用工具规范等\n" +
-                    "3,生活日常\n" +
-                    "家庭情况、饮食作息、消费偏好、出行习惯、休闲爱好等\n" +
-                    "4,健康状况\n" +
-                    "身体症状、慢病困扰、用药习惯、体质特点、就医相关等\n" +
-                    "5,需求偏好\n" +
-                    "高频诉求、内容输出偏好、功能需求、长期规划（理财 / 生活 / 学习）等\n" +
-                    "6,沟通交互\n" +
-                    "说话风格、回复格式偏好、交互习惯、定制化要求等\n" +
-                    "7,关键事件\n" +
-                    "重要时间节点、过往关键经历、待办长期事项、特殊记录等\n" +
-                    "8,知识沉淀\n" +
-                    "高频咨询问题、专属认知观点、常用资料 / 规则、成功处理任务经验等\n" +
+                    "分类1，用户画像\n" +
+                    "内容包含：姓名、昵称、年龄、地域、职业、收入、常用设备、喜好、交流风格、偏好与厌恶、经常提的要求规则等，只记录简短的用户各种标签。\n" +
+                    "整理限制: 请给出一段段简短的用户画像描述作为概要，其他相关内容作为画像内容；用户画像只有汇总出一条完整的画像不需要分割；\n" +
+                    "分类2，任务记录\n" +
+                    "内容包含：正在做的什么事情（开始时间、任务说明、任务过程主要节点、结果、结束时间）\n" +
+                    "整理限制: 每次可以汇总出一条或多条不同任务记录，要根据具体的对话场景和已有的任务记录做判断，那些是旧任务的延续，哪些是新任务的开始；旧任务就更新内容新任务就新增内容；" +
+                    "每条任务都要汇总出一段简短的任务概要；内容要抓住终点，涵盖完整任务内容但不要啰嗦\n" +
                     "========\n" +
-                    "请认真总结记忆得到清单后进行检查，不要有重复的记忆或分类,记忆内容不能胡编乱造信息，要完全从内容中来。\n" +
-                    "在完成记忆总结后，你可以调用记忆操作相关工具，其他未列出的工具都不能用。\n" +
-                    "归类后先保存分类作为父级记忆得到编号，再保存概要内容作为子级记忆，子级记忆的parentId是父级记忆的编号。" ;
+                    "请认真总结记忆得到清单后进行检查，不要有重复的记忆,记忆内容不能胡编乱造信息，要完全从内容中来，冲突的记忆以最新的为准。\n" +
+                    "在完成记忆总结后，你可以调用记忆操作相关工具，其他未列出的工具都不能用。\n"  ;
         }
-        return customPrompt+ "\n本次记忆的agentId是" + agentId;
+        return customPrompt;
     }
 
     public interface MemoryAssistant {
         @UserMessage("{{content}}")
-        String chat(@V("content") String content, InvocationParameters invocationParameters);
+        String chat(@V("content") String content,
+                    ChatRequestParameters chatRequestParameters,
+                    InvocationParameters invocationParameters);
     }
 
-    private String buildMemorySummary(String existingMemory, String newConversation) {
+    /**
+     * @param longTermMemories
+     * @param newConversation
+     * @return
+     */
+    private String buildContent(List<LongTermMemory> longTermMemories, String newConversation) {
         StringBuilder memory = new StringBuilder();
-        memory.append("以下是需要分析的内容");
+        memory.append("以下是需要分析的内容\n");
         memory.append("===========================\n");
-        if (!existingMemory.isEmpty()) {
-            memory.append("【现有记忆】\n").append(existingMemory).append("\n");
+        if(!longTermMemories.isEmpty()){
+            memory.append("以下是现有记忆内容:\n");
+            String memoryContent = longTermMemoryService.buildMemoryContent(longTermMemories);
+            memory.append(memoryContent).append("\n");
         }
         memory.append("【新对话】\n").append(newConversation).append("\n\n");
         memory.append("===========================");
